srcGO: refactored _package-chocolatey CI job for i18n and latest specs
The _package-chocolatey CI job inside srcGO/ directory is completely
outdated and not complying to the latest i18n and specs. Hence, let's
refactor it.

This patch refactors _package-chocolatey CI job for the i18n and latest
specs in srcGO/ directory.

+++ srcGO/.ci/_package-chocolatey_windows-any.ps1
@@ -1,4 +1,4 @@
-# Copyright 2023  (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>
+# Copyright 2023 (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>
 #
 # Licensed under the Apache License, Version 2.0 (the "License"); you may not
 # use this file except in compliance with the License. You may obtain a copy
@@ -15,12 +15,12 @@
 
 # initialize
 if (-not (Test-Path -Path $env:PROJECT_PATH_ROOT)) {
-	Write-Error "[ ERROR ] - Please run from ci.cmd instead!\n"
+	Write-Error "[ ERROR ] - Please run from automataCI\ci.sh.ps1 instead!`n"
 	exit 1
 }
 
-. "${env:PROJECT_PATH_ROOT}\${env:PROJECT_PATH_AUTOMATA}\services\io\os.ps1"
-. "${env:PROJECT_PATH_ROOT}\${env:PROJECT_PATH_AUTOMATA}\services\io\fs.ps1"
+. "${env:LIBS_AUTOMATACI}\services\io\fs.ps1"
+. "${env:LIBS_AUTOMATACI}\services\i18n\translations.ps1"
 
 
 
@@ -36,101 +36,117 @@ function PACKAGE-Assemble-CHOCOLATEY-Content {
 
 
 	# validate project
-	if ($(FS-Is-Target-A-Source "${_target}") -eq 0) {
-		return 10 # not applicable
-	} elseif ($(FS-Is-Target-A-Docs "${_target}") -eq 0) {
-		return 10 # not applicable
-	} elseif ($(FS-Is-Target-A-Library "${_target}") -eq 0) {
-		return 10 # not applicable
-	} elseif ($(FS-Is-Target-A-WASM-JS "${_target}") -eq 0) {
-		return 10 # not applicable
-	} elseif ($(FS-Is-Target-A-WASM "${_target}") -eq 0) {
-		return 10 # not applicable
-	} elseif ($(FS-Is-Target-A-Chocolatey "${_target}") -eq 0) {
-		# accepted
-	} elseif ($(FS-Is-Target-A-Homebrew "${_target}") -eq 0) {
-		return 10 # not applicable
-	} else {
+	if ($(FS-Is-Target-A-Chocolatey "${_target}") -ne 0) {
 		return 10 # not applicable
 	}
 
 
 	# assemble the package
-	$null = FS-Make-Directory "${_directory}\Data\${env:PROJECT_PATH_SOURCE}"
-	$null = FS-Make-Directory "${_directory}\Data\${env:PROJECT_GO}"
-	$null = FS-Make-Directory "${_directory}\Data\automataCI"
-
-	$__process = FS-Copy-All `
-		"${env:PROJECT_PATH_ROOT}\${env:PROJECT_PATH_SOURCE}" `
-		"${_directory}\Data\${env:PROJECT_PATH_SOURCE}"
-	if ($__process -ne 0) {
+	$___source = "${env:PROJECT_PATH_ROOT}\${env:PROJECT_PATH_SOURCE}\"
+	$___dest = "${_directory}\Data\${env:PROJECT_PATH_SOURCE}\"
+	$null = I18N-Assemble "${___source}" "${___dest}"
+	$null = FS-Make-Directory "${___dest}"
+	$___process = FS-Copy-All "${___source}" "${___dest}"
+	if ($___process -ne 0) {
+		$null = I18N-Assemble-Failed
 		return 1
 	}
 
-	$__process = FS-Copy-All `
-		"${env:PROJECT_PATH_ROOT}\${env:PROJECT_GO}" `
-		"${_directory}\Data\${env:PROJECT_GO}"
-	if ($__process -ne 0) {
+	$___source = "${env:PROJECT_PATH_ROOT}\${env:PROJECT_PATH_SOURCE}\.ci\"
+	$___dest = "${_directory}\Data\${env:PROJECT_PATH_SOURCE}\.ci\"
+	$null = I18N-Assemble "${___source}" "${___dest}"
+	$null = FS-Make-Directory "${___dest}"
+	$___process = FS-Copy-All "${___source}" "${___dest}"
+	if ($___process -ne 0) {
+		$null = I18N-Assemble-Failed
 		return 1
 	}
 
-	$__process = FS-Copy-All `
-		"${env:PROJECT_PATH_ROOT}\automataCI" `
-		"${_directory}\Data\automataCI"
-	if ($__process -ne 0) {
+	$___source = "${env:PROJECT_PATH_ROOT}\${env:PROJECT_GO}\"
+	$___dest = "${_directory}\Data\${env:PROJECT_GO}\"
+	$null = I18N-Assemble "${___source}" "${___dest}"
+	$null = FS-Make-Directory "${___dest}"
+	$___process = FS-Copy-All "${___source}" "${___dest}"
+	if ($___process -ne 0) {
+		$null = I18N-Assemble-Failed
 		return 1
 	}
 
-	$__process = FS-Copy-File `
-		"${env:PROJECT_PATH_ROOT}\CONFIG.toml" `
-		"${_directory}\Data"
-	if ($__process -ne 0) {
+	$___source = "${env:PROJECT_PATH_ROOT}\${env:PROJECT_GO}\.ci\"
+	$___dest = "${_directory}\Data\${env:PROJECT_GO}\.ci\"
+	$null = I18N-Assemble "${___source}" "${___dest}"
+	$null = FS-Make-Directory "${___dest}"
+	$___process = FS-Copy-All "${___source}" "${___dest}"
+	if ($___process -ne 0) {
+		$null = I18N-Assemble-Failed
 		return 1
 	}
 
-	$__process = FS-Copy-File `
-		"${env:PROJECT_PATH_ROOT}\ci.cmd" `
-		"${_directory}\Data"
-	if ($__process -ne 0) {
+	$___source = "${env:PROJECT_PATH_ROOT}\automataCI\"
+	$___dest = "${_directory}\Data\automataCI\"
+	$null = I18N-Assemble "${___source}" "${___dest}"
+	$null = FS-Make-Directory "${___dest}"
+	$___process = FS-Copy-All "${___source}" "${___dest}"
+	if ($___process -ne 0) {
+		$null = I18N-Assemble-Failed
 		return 1
 	}
 
-	$__process = FS-Copy-File `
-		"${env:PROJECT_PATH_ROOT}\${env:PROJECT_PATH_SOURCE}\icons\icon-128x128.png" `
-		"${_directory}\icon.png"
-	if ($__process -ne 0) {
+	$___source = "${env:PROJECT_PATH_ROOT}\CONFIG.toml"
+	$___dest = "${_directory}\Data"
+	$null = I18N-Assemble "${___source}" "${___dest}"
+	$___process = FS-Copy-File "${___source}" "${___dest}"
+	if ($___process -ne 0) {
+		$null = I18N-Assemble-Failed
 		return 1
 	}
 
-	$__process = FS-Copy-File `
-		"${env:PROJECT_PATH_ROOT}\README.md" `
-		"${_directory}\README.md"
-	if ($__process -ne 0) {
+	$___source = "${env:PROJECT_PATH_ROOT}\${env:PROJECT_PATH_SOURCE}\icons\icon-128x128.png"
+	$___dest = "${_directory}\icon.png"
+	$null = I18N-Assemble "${___source}" "${___dest}"
+	$___process = FS-Copy-File "${___source}" "${___dest}"
+	if ($___process -ne 0) {
+		$null = I18N-Assemble-Failed
+		return 1
+	}
+
+	$___source = "${env:PROJECT_PATH_ROOT}\README.md"
+	$___dest = "${_directory}\README.md"
+	$null = I18N-Assemble "${___source}" "${___dest}"
+	$___process = FS-Copy-File "${___source}" "${___dest}"
+	if ($___process -ne 0) {
+		$null = I18N-Assemble-Failed
 		return 1
 	}
 
 
 	# REQUIRED: chocolatey required tools\ directory
-	$__process = FS-Make-Directory "${_directory}\tools"
-	if ($__process -ne 0) {
+	$___dest = "${_directory}\tools"
+	$null = I18N-Create "${___dest}"
+	$___process = FS-Make-Directory "${___dest}"
+	if ($___process -ne 0) {
+		$null = I18N-Create-Failed
 		return 1
 	}
 
 
 	# OPTIONAL: chocolatey tools\chocolateyBeforeModify.ps1
-	OS-Print-Status info "scripting tools\chocolateyBeforeModify.ps1..."
-	$__process = FS-Write-File "${_directory}\tools\chocolateyBeforeModify.ps1" @"
+	$___dest = "${_directory}\tools\chocolateyBeforeModify.ps1"
+	$null = I18N-Create "${___dest}"
+	$___process = FS-Write-File "${___dest}" @"
 # REQUIRED - BEGIN EXECUTION
 Write-Host "Performing pre-configurations..."
 "@
-	if ($__process -ne 0) {
+	if ($___process -ne 0) {
+		$null = I18N-Create-Failed
 		return 1
 	}
 
 
 	# REQUIRED: chocolatey tools\chocolateyinstall.ps1
-	OS-Print-Status info "scripting tools\chocolateyinstall.ps1..."
-	$__process = FS-Write-File "${_directory}\tools\chocolateyinstall.ps1" @"
+	$___dest = "${_directory}\tools\chocolateyinstall.ps1"
+	$null = I18N-Create "${___dest}"
+	$___process = FS-Write-File "${___dest}" @"
 # REQUIRED - PREPARING INSTALLATION
 `$tools_dir = "`$(Split-Path -Parent -Path `$MyInvocation.MyCommand.Definition)"
 `$data_dir = "`$(Split-Path -Parent -Path `$tools_dir)\Data"
@@ -144,21 +160,21 @@ Write-Host "Performing pre-configurations..."
 # Materialize the binary
 Write-Host "Building ${env:PROJECT_SKU} (${env:PROJECT_VERSION})..."
 Set-Location "`$data_dir"
-.\ci.cmd setup
+.\automataCI\ci.sh.ps1 setup
 if (`$LASTEXITCODE -ne 0) {
 	Set-Location "`$current_dir"
 	Set-PowerShellExitCode 1
 	return
 }
 
-.\ci.cmd prepare
+.\automataCI\ci.sh.ps1 prepare
 if (`$LASTEXITCODE -ne 0) {
 	Set-Location "`$current_dir"
 	Set-PowerShellExitCode 1
 	return
 }
 
-.\ci.cmd materialize
+.\automataCI\ci.sh.ps1 materialize
 if (`$LASTEXITCODE -ne 0) {
 	Set-Location "`$current_dir"
 	Set-PowerShellExitCode 1
@@ -174,30 +190,38 @@ if (-not (Test-Path "`${data_dir}\bin\${env:PROJECT_SKU}.exe")) {
 
 
 Write-Host "assembling workspace for installation..."
-Move-Item -Path "`${data_dir}\bin" -Destination "`${root_dir}"
-Move-Item -Path "`${data_dir}\lib" -Destination "`${root_dir}"
+if (Test-Path -PathType Container -Path "`${data_dir}\bin") {
+	Move-Item -Path "`${data_dir}\bin" -Destination "`${root_dir}"
+}
+if (Test-Path -PathType Container -Path "`${data_dir}\lib") {
+	Move-Item -Path "`${data_dir}\lib" -Destination "`${root_dir}"
+}
 Set-Location "`$current_dir"
 Remove-Item `$data_dir -Force -Recurse -ErrorAction SilentlyContinue
 "@
-	if ($__process -ne 0) {
+	if ($___process -ne 0) {
+		$null = I18N-Create-Failed
 		return 1
 	}
 
 
 	# REQUIRED: chocolatey tools\chocolateyuninstall.ps1
-	OS-Print-Status info "scripting tools\chocolateyuninstall.ps1..."
-	$__process = FS-Write-File "${_directory}\tools\chocolateyuninstall.ps1" @"
+	$___dest = "${_directory}\tools\chocolateyuninstall.ps1"
+	$null = I18N-Create "${___dest}"
+	$___process = FS-Write-File "${___dest}" @"
 # REQUIRED - PREPARING UNINSTALLATION
 Write-Host "Uninstalling ${env:PROJECT_SKU} (${env:PROJECT_VERSION})..."
 "@
-	if ($__process -ne 0) {
+	if ($___process -ne 0) {
+		$null = I18N-Create-Failed
 		return 1
 	}
 
 
 	# REQUIRED: chocolatey xml.nuspec file
-	OS-Print-Status info "scripting ${env:PROJECT_SKU}.nuspec..."
-	$__process = FS-Write-File "${_directory}\${env:PROJECT_SKU}.nuspec" @"
+	$___dest = "${_directory}\${env:PROJECT_SKU}.nuspec"
+	$null = I18N-Create "${___dest}"
+	$___process = FS-Write-File "${___dest}" @"
 <?xml version="1.0" encoding="utf-8"?>
 <package xmlns="http://schemas.microsoft.com/packaging/2015/06/nuspec.xsd">
 	<metadata>
@@ -214,7 +238,6 @@ Write-Host "Uninstalling ${env:PROJECT_SKU} (${env:PROJECT_VERSION})..."
 	</metadata>
 	<dependencies>
 		<dependency id="chocolatey" version="0.9.8.21" />
-		<dependency id="go" version="1.21.2" />
 	</dependencies>
 	<files>
 		<file src="README.md" target="README.md" />
@@ -224,7 +247,8 @@ Write-Host "Uninstalling ${env:PROJECT_SKU} (${env:PROJECT_VERSION})..."
 	</files>
 </package>
 "@
-	if ($__process -ne 0) {
+	if ($___process -ne 0) {
+		$null = I18N-Create-Failed
 		return 1
 	}
